NewUtilities and NewRed1
NewUtilities uses elevator to scor a glyph. The distances are probably off, but the main structure of the program should be good.
NewRed1 calls NewUtilities and uses the color sensor to detect the jewel (the color sensor code is untested and probably not correct, but we can fix that later).